requirements.txt
selenium ~= 4.22.0
pytest ~= 8.2.2
pytest-html ~= 4.1.1
webdriver-manager ~= 4.0.1
allure-pytest ~= 2.13.5
requests ~= 2.32.3
faker ~= 25.9.2